requirements.txt
async-timeout
attrs
backcall
billiard
biopython
biotite
certifi
charset-normalizer
click
cmake
contourpy
cycler
datasets
debugpy
decorator
dill
executing
fastjsonschema
filelock
fonttools
frozenlist
fsspec
googledrivedownloader
idna
importlib-resources
iprogress
ipykernel
ipywidgets
Jinja2
joblib
kiwisolver
lit
llvmlite
MarkupSafe
matplotlib
mmtf-python
mpmath
msgpack
multidict
nest-asyncio
networkx
numba
numpy==1.23
packaging
pandas<=1.5.3
parso
pexpect
pickleshare
Pillow
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
pyarrow
Pygments
pyparsing
python-dateutil
pytz
PyYAML
pyzmq
quantiprot
regex
requests
ruamel.yaml
ruamel.yaml.clib
safetensors
scikit-learn
scipy
seaborn
six
sympy
threadpoolctl
toml
tornado
tqdm
traitlets
transformers
types-PyYAML
types-toml
typing_extensions
tzdata
urllib3
wcwidth
xxhash
xyzservices
yacs
yarl
zipp

